fix(cms): rename util file name
